@@ -8,7 +8,6 @@
     "@mui/lab": "^5.0.0-alpha.110",
     "@mui/material": "^5.10.16",
     "@mui/styled-engine-sc": "^5.10.16",
-    "@mui/styles": "^5.10.16",
     "@svgdotjs/svg.draggable.js": "^3.0.2",
     "@svgdotjs/svg.js": "^3.1.2",
     "@vitejs/plugin-react": "^2.1.0",
@@ -18,19 +17,19 @@
     "graphql": "^16.6.0",
     "history": "^4.10.1",
     "html-to-image": "^1.11.1",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-color": "^2.19.3",
     "react-cool-dimensions": "^3.0.0",
     "react-copy-to-clipboard": "^5.1.0",
     "react-dnd": "^16.0.1",
     "react-dnd-html5-backend": "^16.0.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-helmet": "^6.1.0",
     "react-router-dom": "^5.2.0",
     "styled-components": "^5.3.6",
     "subscriptions-transport-ws": "^0.11.0",
     "twin.macro": "^3.0.1",
-    "typescript": "^4.5.4",
+    "typescript": "^4.9.3",
     "vite": "^3.1.8",
     "vite-plugin-svgr": "^2.2.2",
     "vite-tsconfig-paths": "^3.5.2",
@@ -61,10 +60,10 @@
     "@graphql-codegen/client-preset": "^1.1.5",
     "@types/history": "^4.7.9",
     "@types/node": "^15.12.2",
-    "@types/react": "^17.0.19",
+    "@types/react": "^18.0.26",
     "@types/react-color": "^3.0.6",
     "@types/react-copy-to-clipboard": "^5.0.0",
-    "@types/react-dom": "^17.0.9",
+    "@types/react-dom": "^18.0.9",
     "@types/react-helmet": "^6.1.5",
     "@types/react-router-dom": "^5.1.8",
     "@typescript-eslint/eslint-plugin": "^5.9.0",
