@@ -1,15 +1,14 @@
 import React from 'react';
 
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 
 import './index.css';
 import App from './App';
 import GlobalStyles from './GlobalStyles';
 
-ReactDOM.render(
+ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
   <>
     <GlobalStyles />
     <App />
   </>,
-  document.getElementById('root'),
 );
